+++ src/main/java/dinkplugin/notifiers/LootNotifier.java
@@ -1,9 +1,6 @@
 package dinkplugin.notifiers;
 
-import com.google.common.cache.Cache;
-import com.google.common.cache.CacheBuilder;
-import com.google.gson.Gson;
-import com.google.gson.JsonSyntaxException;
+import com.google.common.math.DoubleMath;
 import dinkplugin.message.Embed;
 import dinkplugin.message.NotificationBody;
 import dinkplugin.message.NotificationType;
@@ -15,33 +12,34 @@ import dinkplugin.notifiers.data.LootNotificationData;
 import dinkplugin.notifiers.data.SerializedItemStack;
 import dinkplugin.util.ConfigUtil;
 import dinkplugin.util.ItemUtils;
-import dinkplugin.util.SerializedLoot;
+import dinkplugin.util.KillCountService;
+import dinkplugin.util.MathUtils;
+import dinkplugin.util.RarityService;
 import dinkplugin.util.Utils;
 import dinkplugin.util.WorldUtils;
 import lombok.extern.slf4j.Slf4j;
 import net.runelite.api.NPC;
 import net.runelite.api.NpcID;
-import net.runelite.client.config.ConfigManager;
-import net.runelite.client.config.RuneLiteConfig;
 import net.runelite.client.events.NpcLootReceived;
 import net.runelite.client.events.PlayerLootReceived;
 import net.runelite.client.game.ItemManager;
 import net.runelite.client.game.ItemStack;
 import net.runelite.client.plugins.loottracker.LootReceived;
-import net.runelite.client.plugins.loottracker.LootTrackerConfig;
-import net.runelite.client.plugins.loottracker.LootTrackerPlugin;
 import net.runelite.client.util.QuantityFormatter;
 import net.runelite.http.api.loottracker.LootRecordType;
 import org.apache.commons.lang3.StringUtils;
+import org.jetbrains.annotations.Nullable;
 
 import javax.inject.Inject;
 import javax.inject.Singleton;
 import java.util.ArrayList;
 import java.util.Collection;
+import java.util.Comparator;
 import java.util.List;
+import java.util.Map;
+import java.util.Objects;
+import java.util.OptionalDouble;
 import java.util.concurrent.CopyOnWriteArrayList;
-import java.util.concurrent.ScheduledExecutorService;
-import java.util.concurrent.TimeUnit;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 
@@ -49,26 +47,17 @@ import java.util.stream.Collectors;
 @Singleton
 public class LootNotifier extends BaseNotifier {
 
-    private static final String RL_LOOT_PLUGIN_NAME = LootTrackerPlugin.class.getSimpleName().toLowerCase();
-
-    @Inject
-    private Gson gson;
-
     @Inject
     private ItemManager itemManager;
 
     @Inject
-    private ConfigManager configManager;
+    private KillCountService killCountService;
 
     @Inject
-    private ScheduledExecutorService executor;
+    private RarityService rarityService;
 
     private final Collection<Pattern> itemNameAllowlist = new CopyOnWriteArrayList<>();
     private final Collection<Pattern> itemNameDenylist = new CopyOnWriteArrayList<>();
-
-    private final Cache<String, Integer> killCounts = CacheBuilder.newBuilder()
-        .expireAfterAccess(10, TimeUnit.MINUTES)
-        .build();
 
     @Override
     public boolean isEnabled() {
@@ -78,10 +67,6 @@ public class LootNotifier extends BaseNotifier {
     @Override
     protected String getWebhookUrl() {
         return config.lootWebhook();
-    }
-
-    public void reset() {
-        killCounts.invalidateAll();
     }
 
     public void init() {
@@ -118,32 +103,8 @@ public class LootNotifier extends BaseNotifier {
         );
     }
 
-    public void onGameMessage(String message) {
-        // update cached KC via boss chat message with robustness for chat event coming before OR after the loot event
-        KillCountNotifier.parseBoss(message).ifPresent(pair -> {
-            String boss = pair.getKey();
-            Integer kc = pair.getValue();
-
-            // Update cache if no mapping exists (i.e., this is the first boss kill & chat event occurred first)
-            // We store kc - 1 since incrementAndGetKillCount will increment; kc - 1 + 1 == kc
-            Integer existingMapping = killCounts.asMap().putIfAbsent(boss, kc - 1);
-            if (existingMapping != null) {
-                // KC already cached, but we don't know if this boss message appeared before/after the loot event.
-                // If after, we should store kc. If before, we should store kc - 1.
-                // Given this uncertainty, we wait so that the loot event has passed, and then we can store latest kc.
-                executor.schedule(() -> {
-                    killCounts.asMap().merge(boss, kc, Math::max);
-                }, 15, TimeUnit.SECONDS);
-            }
-        });
-    }
-
     public void onNpcLootReceived(NpcLootReceived event) {
-        if (!isEnabled()) {
-            // increment kill count
-            killCounts.asMap().computeIfPresent(event.getNpc().getName(), (k, v) -> v + 1);
-            return;
-        }
+        if (!isEnabled()) return;
 
         NPC npc = event.getNpc();
         int id = npc.getId();
@@ -182,7 +143,7 @@ public class LootNotifier extends BaseNotifier {
     }
 
     private void handleNotify(Collection<ItemStack> items, String dropper, LootRecordType type) {
-        final Integer kc = type == LootRecordType.NPC ? incrementAndGetKillCount(dropper) : null;
+        final Integer kc = killCountService.getKillCount(type, dropper);
         final int minValue = config.minLootValue();
         final boolean icons = config.lootIcons();
 
@@ -213,21 +174,27 @@ public class LootNotifier extends BaseNotifier {
             totalStackValue += totalPrice;
         }
 
-        Evaluable lootMsg;
-        if (!sendMessage && totalStackValue >= minValue && max != null && "Loot Chest".equalsIgnoreCase(dropper)) {
-            // Special case: PK loot keys should trigger notification if total value exceeds configured minimum even
-            // if no single item itself would exceed the min value config - github.com/pajlads/DinkPlugin/issues/403
-            sendMessage = true;
+        Map.Entry<SerializedItemStack, Double> rarest = getRarestDropRate(items, dropper, type);
 
-            JoiningReplacement msg = lootMessage.build();
-            if (msg.getComponents().isEmpty()) {
-                // ensure %LOOT% isn't empty
+        Evaluable lootMsg;
+        if (!sendMessage) {
+            if (sufficientlyRare(rarest)) {
+                // allow notifications for rare drops, even if below configured min loot value
+                sendMessage = true;
+                lootMsg = Replacements.ofMultiple(" ",
+                    Replacements.ofText("Various items including:"),
+                    ItemUtils.templateStack(rarest.getKey(), false)
+                );
+            } else if (totalStackValue >= minValue && max != null && "Loot Chest".equalsIgnoreCase(dropper)) {
+                // Special case: PK loot keys should trigger notification if total value exceeds configured minimum even
+                // if no single item itself would exceed the min value config - github.com/pajlads/DinkPlugin/issues/403
+                sendMessage = true;
                 lootMsg = Replacements.ofMultiple(" ",
                     Replacements.ofText("Various items including:"),
                     ItemUtils.templateStack(max, true)
                 );
             } else {
-                lootMsg = msg;
+                lootMsg = null;
             }
         } else {
             lootMsg = lootMessage.build();
@@ -240,65 +207,57 @@ public class LootNotifier extends BaseNotifier {
                     overrideUrl = config.pkWebhook();
                 }
             }
+            SerializedItemStack keyItem = rarest != null ? rarest.getKey() : max;
+            Double rarity = rarest != null ? rarest.getValue() : null;
             boolean screenshot = config.lootSendImage() && totalStackValue >= config.lootImageMinValue();
+            Evaluable source = type == LootRecordType.PLAYER
+                ? Replacements.ofLink(dropper, config.playerLookupService().getPlayerUrl(dropper))
+                : Replacements.ofWiki(dropper);
             Template notifyMessage = Template.builder()
                 .template(config.lootNotifyMessage())
                 .replacementBoundary("%")
                 .replacement("%USERNAME%", Replacements.ofText(Utils.getPlayerName(client)))
                 .replacement("%LOOT%", lootMsg)
                 .replacement("%TOTAL_VALUE%", Replacements.ofText(QuantityFormatter.quantityToStackSize(totalStackValue)))
-                .replacement("%SOURCE%", Replacements.ofText(dropper))
+                .replacement("%SOURCE%", source)
                 .build();
             createMessage(overrideUrl, screenshot,
                 NotificationBody.builder()
                     .text(notifyMessage)
                     .embeds(embeds)
-                    .extra(new LootNotificationData(serializedItems, dropper, type, kc))
+                    .extra(new LootNotificationData(serializedItems, dropper, type, kc, rarity))
                     .type(NotificationType.LOOT)
-                    .thumbnailUrl(ItemUtils.getItemImageUrl(max.getId()))
+                    .thumbnailUrl(ItemUtils.getItemImageUrl(keyItem.getId()))
                     .build()
             );
         }
     }
 
-    private Integer incrementAndGetKillCount(String npcName) {
-        Integer stored = getStoredKillCount(npcName);
-        if (stored == null) {
-            killCounts.asMap().computeIfPresent(npcName, (k, v) -> v + 1);
-            return null;
-        }
-        return killCounts.asMap().compute(npcName, (npc, cached) -> {
-            if (cached == null || cached < stored) {
-                return stored + 1;
-            }
-            return cached + 1;
-        });
+    @Nullable
+    private Map.Entry<SerializedItemStack, Double> getRarestDropRate(Collection<ItemStack> items, String dropper, LootRecordType type) {
+        if (type != LootRecordType.NPC) return null;
+        return items.stream()
+            .map(item -> {
+                OptionalDouble rarity = rarityService.getRarity(dropper, item.getId(), item.getQuantity());
+                if (rarity.isEmpty()) return null;
+                return Map.entry(item, rarity.getAsDouble());
+            })
+            .filter(Objects::nonNull)
+            .min(Comparator.comparingDouble(Map.Entry::getValue))
+            .map(pair -> {
+                ItemStack item = pair.getKey();
+                SerializedItemStack stack = ItemUtils.stackFromItem(itemManager, item.getId(), item.getQuantity());
+                return Map.entry(stack, pair.getValue());
+            })
+            .orElse(null);
     }
 
-    /**
-     * @param npcName {@link NPC#getName()}
-     * @return the kill count stored by the base runelite loot tracker plugin
-     */
-    private Integer getStoredKillCount(String npcName) {
-        if ("false".equals(configManager.getConfiguration(RuneLiteConfig.GROUP_NAME, RL_LOOT_PLUGIN_NAME))) {
-            // assume stored kc is useless if loot tracker plugin is disabled
-            return null;
-        }
-        String json = configManager.getConfiguration(LootTrackerConfig.GROUP,
-            configManager.getRSProfileKey(),
-            "drops_NPC_" + npcName
-        );
-        if (json == null) {
-            // no kc stored implies first kill
-            return 0;
-        }
-        try {
-            return gson.fromJson(json, SerializedLoot.class).getKills();
-        } catch (JsonSyntaxException e) {
-            // should not occur unless loot tracker changes stored loot pojo structure
-            log.warn("Failed to read kills from loot tracker config", e);
-            return null;
-        }
+    private boolean sufficientlyRare(@Nullable Map.Entry<SerializedItemStack, Double> rarest) {
+        if (rarest == null) return false;
+        int configRareDenominator = config.lootRarityThreshold();
+        if (configRareDenominator <= 0) return false;
+        double rarityThreshold = 1.0 / configRareDenominator;
+        return DoubleMath.fuzzyCompare(rarest.getValue(), rarityThreshold, MathUtils.EPSILON) <= 0;
     }
 
     private static boolean matches(Collection<Pattern> regexps, String input) {