+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 ## Unreleased
 
+- Major: Add item rarity to monster loot and collection notifications. (#425, #427)
+- Minor: Add wiki link for loot source in rich embeds. (#431)
+- Minor: Add chat message upon noteworthy plugin updates. (#429)
+- Minor: Allow loot notifications for rare items below the min value threshold. (#426)
+- Minor: Include relevant kill count in collection log notifications. (#424)
+- Minor: Obtain kill count from chat commands plugin for loot notifications. (#392)
+- Minor: Add chat command to obtain the player's dink hash. (#408)
+- Dev: Refactor kill count tracking from `LootNotifier` to `KillCountService`. (#392)
+
 ## 1.8.4
 
 - Minor: Fire loot notification for PK loot chests with sufficiently high total value. (#417)